Add support for a Needs Sync? checkbox that controls if a Participant is synced
This approach was abandoned in favor of using SIS Imports into Canvas.
Saving this code b/c it has a lot of nice refactoring for the
BZ_ProgramParticipantInfoService.apxc class.

The idea was to have a "Needs Sync?" checkbox on the Participant
that a Process Builder flow (or whatever that newer thing is called)
would set when it detected changes that needed to be synced. A staff
member could also manually set this to force a sync. The Platform sync
was going to only process these Participants.

A big reason why this was a bad approach is b/c the Salesforce trigger
logic to set this would have spanned like 5-6 objects and been hard to maintain
alongside the Platform sync logic.